requirements.txt
openai>=1.0.0
gradio
python-dotenv
#pip install -r requirements.txt